requirements/requirements.txt
Django==1.6.2
Jinja2==2.7.2
MarkupSafe==0.19
Pillow==2.4.0
Pygments==1.6
South==0.8.4
Sphinx==1.2.2
Unipath==1.0
argparse==1.2.1
django-admin-bootstrapped==1.6.4
django-appconf==0.6
django-extensions==1.3.3
django-filter==0.7
django-grappelli==2.5.2
django-guardian==1.2.0
django-imagekit==3.2.1
django-mptt==0.6.0
django-mptt-admin==0.1.8
django-suit==0.2.7
django-tagging==0.3.2
django-thumbs==0.4
docutils==0.11
oauthlib==0.6.1
pilkit==1.1.9
psycopg2==2.5.3
six==1.7.3
wsgiref==0.1.2